pyproject.toml
[tool.poetry]
name = "pyfilechecker"
version = "0.1.0"
description = ""
authors = ["ig-rudenko <ig.rudenko1@yandex.ru>"]

[tool.poetry.dependencies]
python = ">=3.8.10,<4.0.0"
django = "^4.2.3"
celery = "^5.3.1"
flake8 = "^6.0.0"
djangorestframework	= "^3.14.0"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
django-simple-captcha = "^0.5.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

